Using the ergodox now
This changes some keybindings to work better with the ergodox plus the
fact that I use dvorak with it. For exaple, the numbers on the right
half (6-0) now correspond to the tags on the right screen. The layout
changing keys are just adapted for the dvorak layout, so are still on
the same position as they used to.