Source environment from dotfiles dir
Previously I expected most of the files to be linked my homedir which
was fine for the time when I was on my old system which had everything
set up as required. But this is a little unclear and easily forgot on a
new system. Therefor it should just source the file dotfiles dir.