Added script for selecting twitch stream to view
This adds a script which provides a selection of currently live twitch
streams that the given user is following. It them provides these streams
as a dmenu selection which, upon selecting, starts them using
streamlink.

This requires streamlink as well as dmenu. You also need to get and
configure a twitch application to get a client ID you can use.