Move to bspwm from awesomewm
I have moved to bspwm a long time ago, just didn't check in the config
changes yet. The main reason for the change was because the awesomewm v4
release which required a lot of changes in the configuration. However,
this alone was not reason enough to make such a change. The other issue
was that awesomewm, generally, stored everything in it's single rc file
making trying out other setups/config very annoying as it would not only
change the visuals, but also the layouting _and_ keybindings, last of
which was the most annoying. Which it's a totally fine approach, I
started to dislike it more and more. Combined with the changes necessary
in v4 I made the decision to switch.