Avoid making a request for each followed user
When checking if a user is live, the script cycled through all followed
users and made a request for each user. This was very wasteful and
easily hit the request limit. Since the v5 api will soon go away, moving
to helix was a good idea anyway and reduced the amount of requests to 2.
Sadly, this no longer displays the game but that's ok for now.