Improve bspwm start config
This now adds classes for specific kinds of applications that I will
always have on specific desktops after having this setup for a long
time. It now also includes a set of cli applications that I was always
starting manually. The only problem currently is that the nextcloud
client is started so shortly after firefox that opening the confirm page
from it doesn't work; may need to add a sleep for that.